Prefetch consolidation and ENS data for everyone in one go.
Previously we used to fetch ENS and consolidsation data for displays and keys in a loop, one wallet at a time. This was very inefficient and spent a lot of time. This PR changes the logic so that it's done with 2 SQL queries in total and rest is calculated in memory. Works way faster.

Signed-off-by: GelatoGenesis <tarmokalling@gmail.com>